% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ez_crop.R
\name{ez_crop}
\alias{ez_crop}
\title{get cropped raster with extent of another raster}
\usage{
ez_crop(e, z)
}
\arguments{
\item{e}{A raster object.}

\item{z}{a raster or sf object with desired extent}
}
\value{
a raster object.
}
\description{
get cropped raster with extent of another raster
}
