% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_elev.R
\name{get_elev}
\alias{get_elev}
\title{returns cropped lidar or SRTM elevation data, needs upgraded to work on disk for lidar.
Use info in https://stat.ethz.ch/pipermail/r-sig-geo/2020-November/028497.html and GDALinfo() to make version that crops on disk
Until then, just having it in a function will reduce memory use}
\usage{
get_elev(wells, model = "dsm")
}
\arguments{
\item{x}{A raster or sf object.}
}
\value{
raster object.
}
\description{
returns cropped lidar or SRTM elevation data, needs upgraded to work on disk for lidar.
Use info in https://stat.ethz.ch/pipermail/r-sig-geo/2020-November/028497.html and GDALinfo() to make version that crops on disk
Until then, just having it in a function will reduce memory use
}
