% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rgb_array.R
\name{get_rgb_array}
\alias{get_rgb_array}
\title{get a plottable RGB array from a RGB raster}
\usage{
get_rgb_array(ras)
}
\arguments{
\item{x}{A raster object.}
}
\value{
an array object.
}
\description{
get a plottable RGB array from a RGB raster
}
