% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gget_map.R
\name{gget_map}
\alias{gget_map}
\title{get cropped ggmap with extent of another object}
\usage{
gget_map(obj, mt = "satellite")
}
\arguments{
\item{x}{A sf or raster object.}
}
\value{
a gg map object.
}
\description{
get cropped ggmap with extent of another object
}
