% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmap_to_raster.R
\name{ggmap_to_raster}
\alias{ggmap_to_raster}
\title{Convert ggmap to raster}
\usage{
ggmap_to_raster(x)
}
\arguments{
\item{x}{A ggmap object.}
}
\value{
raster object.
}
\description{
Convert ggmap to raster
}
